moved result from pages to components
